requirements.txt
# Pothole Segmentation Project - Requirements
# Install with: pip install -r requirements.txt

# Core ML and Computer Vision
ultralytics>=8.3.0
torch>=1.8.0
torchvision>=0.9.0
opencv-python>=4.6.0
numpy>=1.23.0
pillow>=7.1.2

# Data Processing
pandas>=1.4.0
scipy>=1.4.1

# Visualization
matplotlib>=3.3.0
seaborn>=0.12.0

# Web Interface
streamlit>=1.50.0

# Utilities
pyyaml>=5.3.1
requests>=2.23.0
psutil>=5.8.0

# Optional but recommended
jupyter>=1.0.0
ipywidgets>=8.0.0

